data contract deployer
