+++ packages/data-contract/.env
@@ -1,4 +1,4 @@
 skipSynchronizationBeforeHeight=1000000
-MNEMONIC="your mnemonic"
+MNEMONIC="debate father hello tube crisp gold scatter feel neither remove bar trouble"
 NETWORK="testnet"
-ID=""
+OWNER_IDENTIFIER="7dwjL5frrkM69pv3BsKSQb4ELrMYmDeE11KNoDSefG6c"
